Move FileWrapper to file/
